Split decision_variable.rs into submodules
